+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Flood fill example".
 - Tests for "5D chests".
 - Tests for "Near-palindromes".
+- Tests for "Reversed look-and-say".
 
 ### Fixed
 - Tests groups for "CGFunge interpreter".